auth: Store in home directory instead of in module installation
When installing tourbot from Git and running it from the local
directory via 'npm server', storing the authentication there is fine.

But when users install it via npm-install, we should not store the auth
file in the global /usr/lib/node_modules/mw-tool-tourbot directory.
(especially with a current-user chmod). Use HOME/.config instead.