diff: Factor out simpleDiff from cli, add tests, fix bug
The following example resulted in a misleading diff being shown
to the user:

Removed: "Required: foo, bar"
Added  : "Required: bar"

This was shown to the user as:

> - Required: foo, bar
> + Required:  bar

Note the two spaces. This was because the logic computing diff.textAfter
and diff.removed were iterating over the entire added line until a difference
was found, instead of limiting themselves to the portion already covered
by the first iteration that found diff.textBefore and diff.added.

As such, it was finding the last space before "bar" in "Required: foo, bar" (removed)
as being the same space as in "Required: bar" (added)