cli: Add preloading for next subject while handling subject
As soon as we've got the wikitext of the current page, start
preloading the next in the background.

Even when dealing with a list of pages where most pages have no
matches found (meaning, nono delay from manually staging changes
or saving edits) even then, the preloading makes it skim through
lots of pages much faster.