cli: Exit early if login credentials cache is not valid
Previously, getBotClient (which uses auth) was not called until
inside the handling of individual subjects. Which means, that a
failure from getBotClient will be treated as a handleSubject error,
and thus we continue with the next, and the next and the next.

However it's unlikely that credential validity is specific to
a single wiki.

Workaround this by making getWikiMap() also use getBotClient(),
so that in case of an auth failure there, the wikiMap fails,
which means the process exits early.