cli: Implement 'all' option for proposeChange
This will accept all replacements for the current pattern,
regardless of whether they are exactly the same removed/added line,
and regardless of --auto mode.

For example, when a file has 1000 uses of $j and you trust the
automatic replacement well enough to not want to say yes/no to
each, you can answer with 'a' or 'all', and they will all be
automatically accepted.

The interface will still display each individual replacement diff
one by one, and it will also ask before saving whether you are sure.

This way, you can take time to spot any incorrect replacement and
decide to reject the proposed edits and either do it manually
or e.g. restart the bot to try again without 'all' or something
else.

Given this check at the end, it uses a very short timeout (0.1s)
for each individually auto-accepted diff.