Revert "patterns: Add swap for mw.util.jsMessage>mw.notify"
Too fragile. You can still use tourbot to expand mwgrep output
and automatically map <wiki-id><title> to urls to be opened
in the default browser with "o", which is almost as easy and
a lot less fragile.

tourbot --all --file mwgrep.out

Which will present each file even if there is no diff.