patterns: Add swap for removing "mediawiki.legacy.wikibits" dep
This was meant to be an internal module for base-loading purposes.
It doesn't need to be depended on for wikis that enable legacy
JavaScript (as WMF currently does).

For code that can use dependencies and ResourceLoader (e.g. anything
except legacy user scripts), one should use mw.loader or mw.util
instead.

For anything else, legacy JS will take care of these automatically
without a dependency.

On sites that disable this, it isn't meant to be available via
a dependency either.