patterns: Remove swaps for importScriptURI
This isn't safe due to it having the legacy feature of deduplicating
requests, which some scripts rely on, the very same kinds of scripts
still use this function. For example, some legacy gadgets use it
to declare their dependencies on top of each file relying on it
not loading multiple times and not re-defining object references
etc.

Can still be useful but I basically always skip it because I'm
focussed on a different migration and it just gets in the way.