test: Allow multiple rounds of matching in tests
This is to match what the real CLI does. Previously this part of
the Fixer was missing, which was not needed for any test cases
we have currently, but will be needed in my next commit.

Also drop needless '/g' from replace.js test, which was not needed
and set a bad example since none of the real patterns use the global
flag.