patterns: Add swap for jquery.ui aliases
Including some new magic to attempt better de-duplication
after the swapping. This wasn't done for mediawiki.api previously
but I'm doing it here because it's very common for gadgets and
scripts to depend on several of these at once.

Duplicates are fine in mw.loader, but its nicer when tourbot
doesn't produce such confusing output.

Ref https://phabricator.wikimedia.org/T219604.