content: Refuse editing pages with unclosed <nowiki>
Saving edits to these might trigger `{{subst:` expansion and
completely break the script, even if the submitted syntax
is entirely valid (as we see it).

This does not show in the diff because it happens server-side
while saving, as part of wikitext PST (pre-save transformation).

See <https://phabricator.wikimedia.org/T236828>