patterns: Fix wgEnableAPI swap for `!window.mw.config.get('wgEnableAPI')`
Without this change, the new fixture was changed to the following,
which is incorrect:

    - if ( !window.mw.config.get('wgEnableAPI') )
    + if ( !window.true )

The problem is that [^!] is re-tried at every position, first at the
beginning and then at every next character. For example, at the ' '
before the `!` it matches, but then there is no 'window' or 'mw'
after it. It tries again at the `.` dot after `window` where it matched.
The `.` is not a `!`, and it is followed by `mw.config` without window
again but that is accepted because it is optional.

Avoid this for now by disallowing in general any matches starting from
a dot (property accessor). In this case, the replacement would always
be wrong because window.true does not mean the value "true" it means
a property of which the name is "true".

Ref https://github.com/Krinkle/mw-tool-tourbot/issues/14.