patterns: Remove cleanup of '.charAt()'
This is a matter of coding style preference. As such, I already
demoted it some years ago from a swap to a cleanup (only done if
at least one swap was performed).

But, I skip it almost all the time because it is hard to be sure
in the context of a small diff whether it is safe or not.

There can be any method on any primitive or object called "charAt".
Only when the variable refers to a simple string is it safe to
replace, which is risky, and so I always skipped it.

Remove it properly as it is not deprecated and does not always
improve the script.

I added it originally because many scripts used str.charAt(0)
instead of str[0] because the [0] syntax was not supported
in IE6. So it was part of a clean up in that way, but it is not
needed.