patterns: Fix negated wgIsMainPage swap
When the original gadget code uses a negated match combined with
a namespace check, then we must not remove the whole thing in
favour of just !mainpage. Instead it must be treated the same as
without the namespace match because such conditions are meant
to target "all articles but not the main page".

Whereas for positive matches, the namespace exists for the purpose
of matching the main page itself.

* NS 0 and title=Main Page
  Uses NS to disambiuate from {User/Talk/Template}:Main_Page etc.
  => NS check is redundant if we use wgIsMainPage.

* NS 0 and title != Main Page
  Uses NS to match all articles and then exclude the main page.
  => NS check is NOT redundant even if using wgIsMainPage.