cli: Add `--config` option to change where credentials are stored
This is especially useful if you run tourbot in an ephemeral
container where e.g. only a project directory is mounted and nothing
else persists. In that case, one may want to store the credentials
in whatever your current directory is instead of under the temporary
container's $HOME directory where it would not be retained for next
time.